FCE-1596: Docs restructure (#133)
## Description

There is a lot to read here so I guess this PR will take a long time to
review...
I welcome any of your suggestions and changes. We can treat it as a
draft to work on.

What's new?
- Docs are restructured based on this guide: https://diataxis.fr/
- Much of the existing content just orgnized into a diffrent place
- There is also a lot of new *LLM generated* content. I tried to review
the changes, especially the tutorials and examples if they work and they
do as of 02.07.2025. There are also some new LLM generated pages in
Concepts sections. I tried to make them more verbose and more pleasing
to read and I think this was achived but there might be some LLM
bullshit I might've missed.
- Replaced markdown-spellcheck with https://cspell.org/ because it has
support for MDX. I just couldn't get the old lib to work with mdx
comments, and cspell works great.

---------

Co-authored-by: Tomasz Mazur <47872060+AHGIJMKLKKZNPJKQR@users.noreply.github.com>
Co-authored-by: adrian <33912477+czerwiukk@users.noreply.github.com>
Co-authored-by: Piotr Wodecki <44680063+PiotrWodecki@users.noreply.github.com>