Fce 1897/fix twoslash (#173)
This PR restores Twoslash typechecking.
Previously it was broken due to invalid types between older docs and
newer packages.
Instead of keeping a copy of each SDK for each doc version I opted for
ignoring all errors coming from older docs and only checking the `next`
version.

This is done with a custom remark plugin that injects additional meta
content to each snippet containing the location of the file that the
snippet is contained in. A little hacky, but I didn't see any other
solutions, as Twoslash doesn't seem to get the paths itself.

This is then used in the onError callback that decided on error handling
depending on this meta tag.

Bonus: Switched to the faster build tool chain cutting the build time on
CD from 7.5 min to 5.25 min - I'm noticing much faster local build times
as well